added confirm to deleting a list item, added writefile to app.delete to rewrite after splice, changes id everytime a list is deleted (based on index), needs to be refreshed everytime however to view changes

+++ public/assets/js/index.js
@@ -73,10 +73,16 @@ const handleNoteDelete = function (event) {
     activeNote = {};
   }
 
-  deleteNote(note.id).then(() => {
-    getAndRenderNotes();
-    renderActiveNote();
-  });
+  var yes = confirm("Would you like to delete?");
+
+  if (yes) {
+    deleteNote(note.id).then(() => {
+      getAndRenderNotes();
+      renderActiveNote();
+    });
+  }
+  
+  
 };
 
 // Sets the activeNote and displays it
